+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Master
 * Migrate to lsp-protocol.
+* Implement metals/windowStateDidChange notification.
 
 ## 1.0.0
 * Migrate from `lsp-mode`.